Clean up code to separate picker implementations
