+++ Runtime/Core/SelectionPickerRendererHDRP.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5f130a448e6f0cc45bacc05a4653fc41
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
