+++ Runtime/Core/SelectionPickerRendererURP.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2347b9342bafaa845b0b25387d6fa381
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
